Refactor TrainingLog component and split it into subcomponents
