+++ src/components/training_log/Calendar.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+
+class Calendar extends Component {
+  // TODO
+
+  state = {};
+  render() {
+    return <div />;
+  }
+}
+
+export default Calendar;
